tests: add initial strategy for integration testing
