+++ tests/integration/container/src/constants/actioman_source_container_path.ts
@@ -1,0 +1,7 @@
+/**
+ * Static variable representing the Actioman source directory path inside the container.
+ * @type {URL}
+ */
+export const actiomanSourceContainerPath = new URL(
+  "file:///usr/share/actioman/",
+);
