+++ tests/integration/container/src/constants/app_source_container_path.ts
@@ -1,0 +1,5 @@
+/**
+ * Static variable representing the application source directory path inside the container.
+ * @type {URL}
+ */
+export const appSourceContainerPath = new URL("file:///app/");
