+++ tests/integration/container/src/constants/bun_source_container_path.ts
@@ -1,0 +1,5 @@
+/**
+ * Static variable representing the Bun cache directory path inside the container.
+ * @type {URL}
+ */
+export const bunSourceContainerPath = new URL("file:///root/.bun/");
