+++ tests/integration/container/src/constants/cache_local_path.ts
@@ -1,0 +1,5 @@
+/**
+ * Static variable representing the local cache directory URL for integration tests.
+ * @type {URL}
+ */
+export const cacheLocalPath = new URL("../../../.cache/", import.meta.url);
