+++ tests/integration/container/src/constants/cache_project_local_path.ts
@@ -1,0 +1,7 @@
+import { cacheLocalPath } from "./cache_local_path";
+
+/**
+ * Static variable representing the local project cache directory URL for integration tests.
+ * @type {URL}
+ */
+export const cacheProjectLocalPath = new URL("project/", cacheLocalPath);
