+++ tests/integration/container/src/constants/container_ports.ts
@@ -1,0 +1,20 @@
+/**
+ * Ports exposed by the actioman container for development and testing purposes.
+ *
+ * @property {number} ACTIOMAN Main port for the actioman service
+ * @property {number} TEST Optional port for tests
+ * @property {number} TEST_1 Additional test port 1
+ * @property {number} TEST_2 Additional test port 2
+ * @property {number} TEST_3 Additional test port 3
+ * @property {number} TEST_4 Additional test port 4
+ * @property {number} TEST_5 Additional test port 5
+ */
+export const ACTIOMAN_CONTAINER_PORTS = {
+  ACTIOMAN: 30321, // Puerto del servicio actioman
+  TEST: 30333, // Puerto opcional para tests
+  TEST_1: 30334, // Puerto de prueba adicional 1
+  TEST_2: 30335, // Puerto de prueba adicional 2
+  TEST_3: 30336, // Puerto de prueba adicional 3
+  TEST_4: 30337, // Puerto de prueba adicional 4
+  TEST_5: 30338, // Puerto de prueba adicional 5
+};
