+++ tests/integration/container/src/constants/container_timeout_seconds.ts
@@ -1,0 +1,5 @@
+/**
+ * Static variable for the container timeout in seconds for integration tests.
+ * @type {number}
+ */
+export const CONTAINER_TIMEOUT_SECONDS = /* 10 minutes */ 10 * 60;
