+++ tests/integration/container/src/constants/image_name.ts
@@ -1,0 +1,5 @@
+/**
+ * Static variable for the Docker image name used in integration tests.
+ * @type {string}
+ */
+export const IMAGE_NAME = "oven/bun:latest";
