+++ tests/integration/container/src/constants/project_local_path.ts
@@ -1,0 +1,5 @@
+/**
+ * Static variable representing the local project directory URL for integration tests.
+ * @type {URL}
+ */
+export const projectLocalPath = new URL("../../../../../", import.meta.url);
