+++ tests/integration/container/src/constants/project_source_paths.ts
@@ -1,0 +1,4 @@
+/**
+ * List of project source files and configuration to mount in the container.
+ */
+export const PROJECT_SOURCE_PATHS = ["./src", "./package.json", "./bun.lock"];
