+++ tests/integration/container/src/constants/scripts_local_path.ts
@@ -1,0 +1,5 @@
+/**
+ * Static variable representing the local scripts directory URL for integration tests.
+ * @type {URL}
+ */
+export const scriptsLocalPath = new URL("../../../scripts/", import.meta.url);
