+++ tests/integration/container/src/container_scripts_container_path.ts
@@ -1,0 +1,7 @@
+/**
+ * The absolute path inside the Docker container where scripts are stored.
+ * @type {URL}
+ */
+export const containerScriptsContainerPath = new URL(
+  "file:///root/container_scripts/",
+);
