+++ tests/integration/container/src/container_scripts_local_path.ts
@@ -1,0 +1,5 @@
+/**
+ * The local filesystem path to the directory containing container scripts for integration tests.
+ * @type {URL}
+ */
+export const containerScriptsLocalPath = new URL("../", import.meta.url);
