+++ tests/integration/container/src/dtos/bootstrap_container_options.ts
@@ -1,0 +1,7 @@
+/**
+ * Options for bootstrapping a Docker container in integration tests.
+ * @property {boolean} [verbose] - If true, enables verbose logging during container startup.
+ */
+export type bootstrapContainerOptions = {
+  verbose?: boolean;
+};
