+++ tests/integration/container/src/dtos/kill_container_options.ts
@@ -1,0 +1,7 @@
+/**
+ * Options for killing a Docker container in integration tests.
+ * @property {boolean} [verbose] - If true, enables verbose logging during container shutdown.
+ */
+export type killContainerOptions = {
+  verbose?: boolean;
+};
