+++ tests/integration/container/src/initialize_cli_actioman_options.ts
@@ -1,0 +1,7 @@
+/**
+ * Options for initializing the CLI Actioman process in integration tests.
+ * @property {number} [timeout] - Optional timeout in milliseconds for CLI operations.
+ */
+export type initializeCliActiomanOptions = {
+  timeout?: number;
+};
