+++ tests/integration/container/src/project_cache_paths.ts
@@ -1,0 +1,4 @@
+/**
+ * List of project cache directories to mount in the container.
+ */
+export const PROJECT_CACHE_PATHS = ["./node_modules"];
