--- tests/integration/container_scripts/bootstrap.sh
@@ -1,7 +1,0 @@
-#!/bin/sh
-
-# Sleep for the number of seconds passed as the first argument
-sleep $1 &
-echo $! > /tmp/sleep_pid.tmp
-# Wait for the sleep process to finish
-wait $!
